17.02.

+++ mobile/src/main/java/com/example/eric/application/databaseSource.java
@@ -62,6 +62,20 @@ public class databaseSource{
         return ui_Log;
     }
 
+    //Add user data by use of modified UI_Log
+    public UserInputLog2 create(UserInputLog2 ui_Log2){
+        ContentValues values = new ContentValues();
+        values.put(databaseHelper.USER_ID_V2, ui_Log2.getUser_id());
+        values.put(databaseHelper.VERSUCH_V2, ui_Log2.getVersuch());
+        values.put(databaseHelper.MODALITÄT_V2, ui_Log2.getModalitaet());
+        values.put(databaseHelper.PROZESS_ID_V2, ui_Log2.getProzess_id());
+        values.put(databaseHelper.PROCESS_BLENDIN_V2, ui_Log2.getProcessBlendInTime());
+        values.put(databaseHelper.CLICKTIME, ui_Log2.getConfirmationtime());
+        database.insert(databaseHelper.TABLENAME_V2, null, values);
+        Log.e("Database Source", "ONE ROW INSERTED...");
+        return ui_Log2;
+    }
+
     //close database, here for logcat
     public void close(){
         if (database != null) {
@@ -71,10 +85,18 @@ public class databaseSource{
     }
 
     //deletes data of a user
-    public void deleteData(long User) {
-        String whereClause =
-                databaseHelper.USER_ID + " = " + User;
-        database.delete(databaseHelper.TABLENAME, whereClause, null);
+    public void deleteData(long User, String modalitaet, String Tablename) {
+       String whereClause;
+        if(Tablename==databaseHelper.TABLENAME) {
+            whereClause = databaseHelper.USER_ID + " = " + User
+                    + " AND " + databaseHelper.MODALITÄT + " = " + modalitaet;
+            database.delete(databaseHelper.TABLENAME, whereClause, null);
+        }
+        if(Tablename==databaseHelper.TABLENAME_V2) {
+            whereClause = databaseHelper.USER_ID + " = " + User
+                    + " AND " + databaseHelper.MODALITÄT_V2 + " = " + modalitaet;
+            database.delete(databaseHelper.TABLENAME_V2, whereClause, null);
+        }
     }
 
     //open database
